same

+++ code/mainhpc.jl
@@ -27,8 +27,8 @@ InfectiousProfile = "empirical";
 MeanOfLognormal = getMeanOfLognormalDistribution();
 
 # Societal details
-WaitBeforeTestTaken  = parse(Int, ARGS[1]);  # Number of days before test is taken
-WaitBeforeTestResult  = parse(Int, ARGS[2]); # Number of days before test result arrives after test is taken
+WaitBeforeTestTaken  = parse(Int64, ARGS[1]);  # Number of days before test is taken
+WaitBeforeTestResult  = parse(Int64, ARGS[2]); # Number of days before test result arrives after test is taken
 
 # Test-and-trace details
 ProbabilityChildIsTraced  = -0.02; #+34*0.02 // Fraction of children that are found through contact tracing
